Move owner processes to a supervisor
